:truck: :: Init repository naming
