+++ Service/Sources/Domain/UseCases/Club/FetchManagementClubUseCase.swift
@@ -2,8 +2,8 @@ import RxSwift
 import Foundation
 
 public final class FetchManagementClubUseCase {
-    public init(repository: ClubRepository) {
-        self.repository = repository
+    public init(clubRepository: ClubRepository) {
+        self.repository = clubRepository
     }
     
     private let repository: ClubRepository
